tests/wpt/web-platform-tests/2dcontext/drawing-rectangles-to-the-canvas/2d.clearRect.negative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.clearRect.negative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.clearRect.negative</h1>
<p class="desc">clearRect of negative sizes works</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/clear-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("clearRect of negative sizes works");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.clearRect(0, 0, 50, 25);
ctx.clearRect(100, 0, -50, 25);
ctx.clearRect(0, 50, 50, -25);
ctx.clearRect(100, 50, -50, -25);
_assertPixel(canvas, 25,12, 0,0,0,0, "25,12", "0,0,0,0");
_assertPixel(canvas, 75,12, 0,0,0,0, "75,12", "0,0,0,0");
_assertPixel(canvas, 25,37, 0,0,0,0, "25,37", "0,0,0,0");
_assertPixel(canvas, 75,37, 0,0,0,0, "75,37", "0,0,0,0");


});
</script>
